requirements.txt
openai
opencv-python-headless
numpy
scipy
fastapi
uvicorn
pydantic
python-dotenv
Pillow
scikit-learn